Add a Sink type unifying data ...err... sinks. Start on JsonWriters for Sinks. Start on a BlueEyes server.
The Sink type defines a + method. Sinks could be Monoids but they are mostly destructive and I can't see what we'd get from from this abstraction.

JsonWriters are obvious.

Switch from Spray to BlueEyes to get back some type checking. The server responds to two paths: /event to add an event, and /view to get back data. JSON in and out. All requests must include a key field in their JSON content, which is used to dispatch the content.